refactor(libs): integrate types directory into libs
