+++ src/libs/common/time-operation/get-time.ts
@@ -1,4 +1,4 @@
-import { ISOString, LocaleString, Timestamp } from '../../../types';
+import { ISOString, LocaleString, Timestamp } from '../types';
 
 export const getTime = () => Date.now() as Timestamp;
 export const getISOString = () => new Date().toISOString() as ISOString;
